requirements.txt
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
hydra-core
omegaconf
scikit-learn
numpy
